move type definitions to converters
Admittedly it felt wrong when I imported from process into the converters. Turns
out this will trigger configuration of logging and mess with other libraries.
